+++ core/src/main/java/module-info.java
@@ -252,6 +252,8 @@ module ai.timefold.solver.core {
     opens ai.timefold.solver.core.config.solver.termination to jakarta.xml.bind, org.glassfish.jaxb.runtime;
     opens ai.timefold.solver.core.config.util to jakarta.xml.bind, org.glassfish.jaxb.runtime;
 
+    exports ai.timefold.solver.core.impl.move.builtin;
+
     requires jakarta.xml.bind;
     requires java.xml;
     requires micrometer.core;
