+++ core/src/test/java/ai/timefold/solver/core/impl/move/builtin/ListSwapMoveTest.java
@@ -1,7 +1,8 @@
-package ai.timefold.solver.core.preview.api.move.builtin;
+package ai.timefold.solver.core.impl.move.builtin;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import ai.timefold.solver.core.preview.api.move.builtin.Moves;
 import ai.timefold.solver.core.preview.api.move.test.MoveTester;
 import ai.timefold.solver.core.testdomain.list.TestdataListEntity;
 import ai.timefold.solver.core.testdomain.list.TestdataListSolution;
