+++ core/src/test/java/ai/timefold/solver/core/impl/move/builtin/SampleValueRangesBiasIT.java
@@ -1,4 +1,4 @@
-package ai.timefold.solver.core.preview.api.move.builtin;
+package ai.timefold.solver.core.impl.move.builtin;
 
 import java.util.List;
 import java.util.Random;
