+++ core/src/test/java/ai/timefold/solver/core/impl/move/builtin/SubListChangeMoveTest.java
@@ -1,10 +1,12 @@
-package ai.timefold.solver.core.preview.api.move.builtin;
+package ai.timefold.solver.core.impl.move.builtin;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
 
 import ai.timefold.solver.core.api.domain.common.Lookup;
+import ai.timefold.solver.core.preview.api.domain.metamodel.ElementPosition;
+import ai.timefold.solver.core.preview.api.move.builtin.Moves;
 import ai.timefold.solver.core.preview.api.move.test.MoveTester;
 import ai.timefold.solver.core.preview.api.neighborhood.stream.dataset.sample.Range;
 import ai.timefold.solver.core.testdomain.list.TestdataListEntity;
@@ -14,71 +16,68 @@ import ai.timefold.solver.core.testdomain.list.TestdataListValue;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
-class SubListSwapMoveTest {
+class SubListChangeMoveTest {
 
     @Nested
-    class SwapWithinSameEntity {
+    class MoveWithinSameEntity {
 
         @Test
-        void swapNonOverlappingSpans() {
+        void moveSpanForward() {
             var value1 = new TestdataListValue("A");
             var value2 = new TestdataListValue("B");
             var value3 = new TestdataListValue("C");
             var value4 = new TestdataListValue("D");
             var value5 = new TestdataListValue("E");
-            var value6 = new TestdataListValue("F");
-            var entity = new TestdataListEntity("Entity", value1, value2, value3, value4, value5, value6);
+            var entity = new TestdataListEntity("Entity", value1, value2, value3, value4, value5);
             var solution = new TestdataListSolution();
             solution.setEntityList(List.of(entity));
-            solution.setValueList(List.of(value1, value2, value3, value4, value5, value6));
+            solution.setValueList(List.of(value1, value2, value3, value4, value5));
 
             var solutionMetaModel = TestdataListSolution.buildMetaModel();
             var variableMetaModel = solutionMetaModel.genuineEntity(TestdataListEntity.class)
                     .listVariable("valueList", TestdataListValue.class);
 
-            var move = Moves.swap(variableMetaModel, new Range<>(entity, 1, 3), new Range<>(entity, 4, 6), false);
+            var move = Moves.change(variableMetaModel, new Range<>(entity, 1, 3), ElementPosition.of(entity, 3), false);
 
             MoveTester.build(solutionMetaModel)
                     .using(solution)
                     .execute(move);
 
-            assertThat(entity.getValueList()).containsExactly(value1, value5, value6, value4, value2, value3);
+            assertThat(entity.getValueList()).containsExactly(value1, value4, value5, value2, value3);
         }
 
         @Test
-        void swapNormalizesOutOfOrderSpans() {
-            // The same swap as swapNonOverlappingSpans, but with left/right passed in reverse order.
+        void moveSpanBackwardReversing() {
             var value1 = new TestdataListValue("A");
             var value2 = new TestdataListValue("B");
             var value3 = new TestdataListValue("C");
             var value4 = new TestdataListValue("D");
             var value5 = new TestdataListValue("E");
-            var value6 = new TestdataListValue("F");
-            var entity = new TestdataListEntity("Entity", value1, value2, value3, value4, value5, value6);
+            var entity = new TestdataListEntity("Entity", value1, value2, value3, value4, value5);
             var solution = new TestdataListSolution();
             solution.setEntityList(List.of(entity));
-            solution.setValueList(List.of(value1, value2, value3, value4, value5, value6));
+            solution.setValueList(List.of(value1, value2, value3, value4, value5));
 
             var solutionMetaModel = TestdataListSolution.buildMetaModel();
             var variableMetaModel = solutionMetaModel.genuineEntity(TestdataListEntity.class)
                     .listVariable("valueList", TestdataListValue.class);
 
-            var move = Moves.swap(variableMetaModel, new Range<>(entity, 4, 6), new Range<>(entity, 1, 3), false);
+            var move = Moves.change(variableMetaModel, new Range<>(entity, 3, 5), ElementPosition.of(entity, 0), true);
 
             MoveTester.build(solutionMetaModel)
                     .using(solution)
                     .execute(move);
 
-            assertThat(entity.getValueList()).containsExactly(value1, value5, value6, value4, value2, value3);
+            assertThat(entity.getValueList()).containsExactly(value5, value4, value1, value2, value3);
         }
 
     }
 
     @Nested
-    class SwapBetweenEntities {
+    class MoveBetweenEntities {
 
         @Test
-        void swapSpansOfDifferingLength() {
+        void moveSpanToOtherEntity() {
             var valueA1 = new TestdataListValue("A1");
             var valueA2 = new TestdataListValue("A2");
             var valueA3 = new TestdataListValue("A3");
@@ -96,14 +95,38 @@ class SubListSwapMoveTest {
             var variableMetaModel = solutionMetaModel.genuineEntity(TestdataListEntity.class)
                     .listVariable("valueList", TestdataListValue.class);
 
-            var move = Moves.swap(variableMetaModel, new Range<>(entityA, 1, 3), new Range<>(entityB, 0, 1), false);
+            var move =
+                    Moves.change(variableMetaModel, new Range<>(entityA, 1, 3), ElementPosition.of(entityB, 1), false);
 
             MoveTester.build(solutionMetaModel)
                     .using(solution)
                     .execute(move);
 
-            assertThat(entityA.getValueList()).containsExactly(valueA1, valueB1, valueA4);
-            assertThat(entityB.getValueList()).containsExactly(valueA2, valueA3, valueB2, valueB3);
+            assertThat(entityA.getValueList()).containsExactly(valueA1, valueA4);
+            assertThat(entityB.getValueList()).containsExactly(valueB1, valueA2, valueA3, valueB2, valueB3);
+        }
+
+        @Test
+        void moveSpanToUnassignedIsSubListUnassignMove() {
+            var value1 = new TestdataListValue("A");
+            var value2 = new TestdataListValue("B");
+            var value3 = new TestdataListValue("C");
+            var entity = new TestdataListEntity("Entity", value1, value2, value3);
+            var solution = new TestdataListSolution();
+            solution.setEntityList(List.of(entity));
+            solution.setValueList(List.of(value1, value2, value3));
+
+            var solutionMetaModel = TestdataListSolution.buildMetaModel();
+            var variableMetaModel = solutionMetaModel.genuineEntity(TestdataListEntity.class)
+                    .listVariable("valueList", TestdataListValue.class);
+
+            var move = Moves.unassign(variableMetaModel, new Range<>(entity, 0, 2));
+
+            MoveTester.build(solutionMetaModel)
+                    .using(solution)
+                    .execute(move);
+
+            assertThat(entity.getValueList()).containsExactly(value3);
         }
 
     }
@@ -116,24 +139,25 @@ class SubListSwapMoveTest {
             var value1 = new TestdataListValue("A");
             var value2 = new TestdataListValue("B");
             var value3 = new TestdataListValue("C");
-            var value4 = new TestdataListValue("D");
-            var entity = new TestdataListEntity("Entity", value1, value2, value3, value4);
+            var entity = new TestdataListEntity("Entity", value1, value2, value3);
             var solution = new TestdataListSolution();
             solution.setEntityList(List.of(entity));
-            solution.setValueList(List.of(value1, value2, value3, value4));
+            solution.setValueList(List.of(value1, value2, value3));
 
             var solutionMetaModel = TestdataListSolution.buildMetaModel();
             var variableMetaModel = solutionMetaModel.genuineEntity(TestdataListEntity.class)
                     .listVariable("valueList", TestdataListValue.class);
 
-            var move = Moves.swap(variableMetaModel, new Range<>(entity, 0, 1), new Range<>(entity, 3, 4), false);
+            // destinationIndex is post-removal: after removing [A, B], the list is [C] (size 1), so index 1 appends.
+            var move = Moves.change(variableMetaModel, new Range<>(entity, 0, 2), ElementPosition.of(entity, 1), false);
 
             MoveTester.build(solutionMetaModel)
                     .using(solution)
-                    .executeTemporarily(move,
-                            view -> assertThat(entity.getValueList()).containsExactly(value4, value2, value3, value1));
+                    .executeTemporarily(move, view -> assertThat(entity.getValueList()).containsExactly(value3, value1,
+                            value2));
 
-            assertThat(entity.getValueList()).containsExactly(value1, value2, value3, value4);
+            // Undone automatically after the callback returns.
+            assertThat(entity.getValueList()).containsExactly(value1, value2, value3);
         }
 
     }
@@ -142,34 +166,48 @@ class SubListSwapMoveTest {
     class MoveProperties {
 
         @Test
-        void getPlanningEntitiesForDifferentEntitiesSwap() {
+        void getPlanningEntitiesForSameEntityMove() {
+            var entity = new TestdataListEntity("Entity", new TestdataListValue("A"), new TestdataListValue("B"));
+            var solutionMetaModel = TestdataListSolution.buildMetaModel();
+            var variableMetaModel = solutionMetaModel.genuineEntity(TestdataListEntity.class)
+                    .listVariable("valueList", TestdataListValue.class);
+
+            var move = (SubListChangeMove<TestdataListSolution, TestdataListEntity, TestdataListValue>) Moves.change(
+                    variableMetaModel, new Range<>(entity, 0, 1), ElementPosition.of(entity, 2), false);
+
+            assertThat(move.getPlanningEntities()).containsExactly(entity);
+        }
+
+        @Test
+        void getPlanningEntitiesForDifferentEntitiesMove() {
             var entityA = new TestdataListEntity("EntityA", new TestdataListValue("A"));
             var entityB = new TestdataListEntity("EntityB", new TestdataListValue("B"));
             var solutionMetaModel = TestdataListSolution.buildMetaModel();
             var variableMetaModel = solutionMetaModel.genuineEntity(TestdataListEntity.class)
                     .listVariable("valueList", TestdataListValue.class);
 
-            var move = (SubListSwapMove<TestdataListSolution, TestdataListEntity, TestdataListValue>) Moves.swap(
-                    variableMetaModel, new Range<>(entityA, 0, 1), new Range<>(entityB, 0, 1), false);
+            var move = (SubListChangeMove<TestdataListSolution, TestdataListEntity, TestdataListValue>) Moves.change(
+                    variableMetaModel, new Range<>(entityA, 0, 1), ElementPosition.of(entityB, 0), false);
 
             assertThat(move.getPlanningEntities()).containsExactlyInAnyOrder(entityA, entityB);
         }
 
         @Test
-        void getPlanningValuesReturnsPreMoveValues() {
+        void getPlanningValuesReturnsMovedValues() {
             var value1 = new TestdataListValue("A");
             var value2 = new TestdataListValue("B");
-            var entity = new TestdataListEntity("Entity", value1, value2);
+            var value3 = new TestdataListValue("C");
+            var entity = new TestdataListEntity("Entity", value1, value2, value3);
             var solution = new TestdataListSolution();
             solution.setEntityList(List.of(entity));
-            solution.setValueList(List.of(value1, value2));
+            solution.setValueList(List.of(value1, value2, value3));
 
             var solutionMetaModel = TestdataListSolution.buildMetaModel();
             var variableMetaModel = solutionMetaModel.genuineEntity(TestdataListEntity.class)
                     .listVariable("valueList", TestdataListValue.class);
 
-            var move = (SubListSwapMove<TestdataListSolution, TestdataListEntity, TestdataListValue>) Moves.swap(
-                    variableMetaModel, new Range<>(entity, 0, 1), new Range<>(entity, 1, 2), false);
+            var move = (SubListChangeMove<TestdataListSolution, TestdataListEntity, TestdataListValue>) Moves.change(
+                    variableMetaModel, new Range<>(entity, 0, 2), ElementPosition.of(entity, 1), false);
 
             MoveTester.build(solutionMetaModel)
                     .using(solution)
@@ -186,9 +224,9 @@ class SubListSwapMoveTest {
             var variableMetaModel = solutionMetaModel.genuineEntity(TestdataListEntity.class)
                     .listVariable("valueList", TestdataListValue.class);
 
-            var move1 = Moves.swap(variableMetaModel, new Range<>(entityA, 0, 1), new Range<>(entityB, 0, 1), false);
-            var move2 = Moves.swap(variableMetaModel, new Range<>(entityA, 0, 1), new Range<>(entityB, 0, 1), false);
-            var move3 = Moves.swap(variableMetaModel, new Range<>(entityA, 0, 1), new Range<>(entityB, 0, 1), true);
+            var move1 = Moves.change(variableMetaModel, new Range<>(entityA, 0, 1), ElementPosition.of(entityB, 0), false);
+            var move2 = Moves.change(variableMetaModel, new Range<>(entityA, 0, 1), ElementPosition.of(entityB, 0), false);
+            var move3 = Moves.change(variableMetaModel, new Range<>(entityA, 0, 1), ElementPosition.of(entityB, 0), true);
 
             assertThat(move1).isEqualTo(move2);
             assertThat(move1.hashCode()).isEqualTo(move2.hashCode());
@@ -199,13 +237,13 @@ class SubListSwapMoveTest {
 
         @Test
         void toStringContainsMoveDetails() {
-            var entityA = new TestdataListEntity("EntityA", new TestdataListValue("A"));
-            var entityB = new TestdataListEntity("EntityB", new TestdataListValue("B"));
+            var entityA = new TestdataListEntity("EntityA", new TestdataListValue("A"), new TestdataListValue("B"));
+            var entityB = new TestdataListEntity("EntityB", new TestdataListValue("C"));
             var solutionMetaModel = TestdataListSolution.buildMetaModel();
             var variableMetaModel = solutionMetaModel.genuineEntity(TestdataListEntity.class)
                     .listVariable("valueList", TestdataListValue.class);
 
-            var move = Moves.swap(variableMetaModel, new Range<>(entityA, 0, 1), new Range<>(entityB, 0, 1), true);
+            var move = Moves.change(variableMetaModel, new Range<>(entityA, 0, 2), ElementPosition.of(entityB, 1), true);
 
             var toString = move.toString();
             assertThat(toString).contains("EntityA").contains("EntityB").contains("reversing-");
@@ -227,8 +265,8 @@ class SubListSwapMoveTest {
             var variableMetaModel = solutionMetaModel.genuineEntity(TestdataListEntity.class)
                     .listVariable("valueList", TestdataListValue.class);
 
-            var originalMove = (SubListSwapMove<TestdataListSolution, TestdataListEntity, TestdataListValue>) Moves
-                    .swap(variableMetaModel, new Range<>(entityA, 0, 1), new Range<>(entityB, 0, 1), false);
+            var originalMove = (SubListChangeMove<TestdataListSolution, TestdataListEntity, TestdataListValue>) Moves
+                    .change(variableMetaModel, new Range<>(entityA, 0, 1), ElementPosition.of(entityB, 0), false);
 
             var rebasedMove = originalMove.rebase(new Lookup() {
                 @Override
@@ -243,8 +281,8 @@ class SubListSwapMoveTest {
                 }
             });
 
-            assertThat(rebasedMove.getLeftRange().entity()).isEqualTo(rebasedEntityA);
-            assertThat(rebasedMove.getRightRange().entity()).isEqualTo(rebasedEntityB);
+            assertThat(rebasedMove.getSource().entity()).isEqualTo(rebasedEntityA);
+            assertThat(rebasedMove.getDestination().<TestdataListEntity> entity()).isEqualTo(rebasedEntityB);
         }
 
     }
