+++ core/src/test/java/ai/timefold/solver/core/impl/move/builtin/SubListUnassignMoveTest.java
@@ -1,10 +1,11 @@
-package ai.timefold.solver.core.preview.api.move.builtin;
+package ai.timefold.solver.core.impl.move.builtin;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
 
 import ai.timefold.solver.core.api.domain.common.Lookup;
+import ai.timefold.solver.core.preview.api.move.builtin.Moves;
 import ai.timefold.solver.core.preview.api.move.test.MoveTester;
 import ai.timefold.solver.core.preview.api.neighborhood.stream.dataset.sample.Range;
 import ai.timefold.solver.core.testdomain.list.unassignedvar.TestdataAllowsUnassignedValuesListEntity;
