+++ core/src/test/java/ai/timefold/solver/core/preview/api/move/builtin/TwoOptListMoveProviderTest.java
@@ -6,6 +6,8 @@ import java.util.ArrayList;
 import java.util.List;
 
 import ai.timefold.solver.core.api.solver.SolutionManager;
+import ai.timefold.solver.core.impl.move.builtin.SubListChangeMove;
+import ai.timefold.solver.core.impl.move.builtin.SubListSwapMove;
 import ai.timefold.solver.core.preview.api.move.test.MoveTester;
 import ai.timefold.solver.core.preview.api.neighborhood.stream.dataset.sample.Range;
 import ai.timefold.solver.core.preview.api.neighborhood.test.NeighborhoodTester;
